+++ gp-includes/things/glossary-entry.php
@@ -65,6 +65,7 @@ class GP_Glossary_Entry extends GP_Thing {
 			'preposition'  => _x( 'preposition', 'part-of-speech', 'glotpress' ),
 			'pronoun'      => _x( 'pronoun', 'part-of-speech', 'glotpress' ),
 			'expression'   => _x( 'expression', 'part-of-speech', 'glotpress' ),
+			'abbreviation' => _x( 'abbreviation', 'part-of-speech', 'glotpress' ),
 		);
 	}
 
@@ -78,6 +79,7 @@ class GP_Glossary_Entry extends GP_Thing {
 	public function restrict_fields( $rules ) {
 		$rules->term_should_not_be( 'empty' );
 		$rules->part_of_speech_should_not_be( 'empty' );
+		$rules->part_of_speech_should_be( 'one_of', $this->parts_of_speech );
 		$rules->glossary_id_should_be( 'positive_int' );
 		$rules->last_edited_by_should_be( 'positive_int' );
 	}