+++ CHANGELOG.md
@@ -12,10 +12,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Integrate three new translation warnings for mismatched URLs, missing named placeholders, and unexpected sprintf tokens.
   Also improve the check for HTML tags. ([#1243](https://github.com/GlotPress/GlotPress-WP/pull/1243))
 * Locales: Add definition for Western Balochi. ([#1254](https://github.com/GlotPress/GlotPress-WP/issues/1254))
+* Add 'abbreviation' as new part of speech value for glossary entries. ([#1199](https://github.com/GlotPress/GlotPress-WP/issues/1199))
 
 **Bugfixes**
 
 * Allow to use the same glossary term and translation for a different part of speech. ([#1251](https://github.com/GlotPress/GlotPress-WP/pull/1251))
+* Validate part of speech value for a glossary entry before inserting into database. ([#1200](https://github.com/GlotPress/GlotPress-WP/issues/1200))
 
 ## [3.0.0-alpha.3] (April 29, 2021)